Add check for page-errors folder

+++ admin/helpers/amazon/clearPageErrors.js
@@ -2,5 +2,8 @@ const path = require("path");
 const fs = require("fs-extra");
 const pageErrorsDir = path.resolve(__dirname, "page-errors");
 
-module.exports = () =>
-  fs.rmdirSync(pageErrorsDir, { force: true, recursive: true });
+module.exports = () => {
+  if ( fs.existsSync(pageErrorsDir) ) {
+    fs.rmdirSync(pageErrorsDir, { force: true, recursive: true });
+  }
+}
